requirements.txt
fastapi
uvicorn
langchain
langchain_groq
langserve
python-decouple
PyPDF2